+++ src/TestSuite/TestCase.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * This file is part of cakephp-link-scanner.
  *
@@ -35,7 +36,7 @@ abstract class TestCase extends BaseTestCase
 
         @unlink(TMP . 'cakephp-link-scanner' . DS . 'link_scanner_lock_file');
 
-        Cache::clearAll();
+        Cache::clear('LinkScanner');
     }
 
     /**
@@ -115,8 +116,7 @@ abstract class TestCase extends BaseTestCase
      */
     protected function getEventManager(LinkScanner $LinkScanner = null)
     {
-        $LinkScanner = $LinkScanner ?: $this->LinkScanner;
-        $eventManager = $LinkScanner->getEventManager();
+        $eventManager = ($LinkScanner ?: $this->LinkScanner)->getEventManager();
 
         if (!$eventManager->getEventList()) {
             $eventManager->setEventList(new EventList());
@@ -136,11 +136,10 @@ abstract class TestCase extends BaseTestCase
      */
     protected function getResponseWithBody($body, Response $response = null)
     {
-        $response = $response ?: new Response();
         $stream = new Stream('php://memory', 'wb+');
         $stream->write($body);
         $stream->rewind();
 
-        return $response->withBody($stream);
+        return ($response ?: new Response())->withBody($stream);
     }
 }