+++ src/Utility/LinkScanner.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * This file is part of cakephp-link-scanner.
  *
@@ -57,6 +58,7 @@ class LinkScanner implements Serializable
     protected $_defaultConfig = [
         'cache' => true,
         'excludeLinks' => '/[\{\}+]/',
+        'exportOnlyBadResults' => false,
         'externalLinks' => true,
         'followRedirects' => false,
         'fullBaseUrl' => null,
@@ -102,9 +104,12 @@ class LinkScanner implements Serializable
     protected $startTime = 0;
 
     /**
-     * Construct
-     * @param \Cake\Http\Client|null $Client A Client instance or null
-     * @param \LinkScanner\ResultScan|null $ResultScan A Client instance or null
+     * Construct.
+     *
+     * Creates `Client` and `ResultScan` instances and loads a possible
+     *  configuration file.
+     * @param \Cake\Http\Client|null $Client A Client instance
+     * @param \LinkScanner\ResultScan|null $ResultScan A ResultScan instance
      * @uses $Client
      * @uses $ResultScan
      */
@@ -115,9 +120,7 @@ class LinkScanner implements Serializable
 
         if (file_exists(CONFIG . 'link_scanner.php')) {
             $config = (new PhpConfig())->read('link_scanner');
-            if (isset($config['LinkScanner'])) {
-                $this->setConfig($config['LinkScanner']);
-            }
+            $this->setConfig($config['LinkScanner'] ?? []);
         }
     }
 
@@ -129,6 +132,34 @@ class LinkScanner implements Serializable
     public function __get($name)
     {
         return $this->$name;
+    }
+
+    /**
+     * Internal method to check if an url can be scanned.
+     *
+     * Returns `false` if:
+     *  - the url has already been scanned;
+     *  - it's an external url and the external url scan has been disabled;
+     *  - the url matches the url patterns to be excluded.
+     * @param string $url Url to check
+     * @return bool
+     * @uses $alreadyScanned
+     * @uses $hostname
+     */
+    protected function _canBeScanned(string $url)
+    {
+        if (!is_url($url) || in_array($url, $this->alreadyScanned) ||
+            (!$this->getConfig('externalLinks') && is_external_url($url, $this->hostname))) {
+            return false;
+        }
+
+        foreach ((array)$this->getConfig('excludeLinks') as $pattern) {
+            if (preg_match($pattern, $url)) {
+                return false;
+            }
+        }
+
+        return true;
     }
 
     /**
@@ -162,9 +193,11 @@ class LinkScanner implements Serializable
     }
 
     /**
-     * Performs a single GET request and returns a `ScanResponse` instance.
+     * Internal method to perform a single GET request and return a
+     *  `Response` instance.
      *
-     * The response will be cached, if that's ok and the cache is enabled.
+     * The response will be cached, if the cache is enabled and if the repsonse
+     *  is "ok" or a redirect.
      * @param string $url The url or path you want to request
      * @return \Cake\Http\Client\Response
      * @uses $Client
@@ -176,27 +209,26 @@ class LinkScanner implements Serializable
         $this->alreadyScanned[] = $url;
         $cacheKey = sprintf('response_%s', md5(serialize($url)));
 
+        //Tries to get the response from the cache
         $response = $this->getConfig('cache') ? Cache::read($cacheKey, 'LinkScanner') : null;
-
         if ($response && is_array($response)) {
             list($response, $body) = $response;
 
             $stream = new Stream('php://memory', 'wb+');
             $stream->write($body);
             $stream->rewind();
-            $response = $response->withBody($stream);
+
+            return $response->withBody($stream);
         }
 
-        if (!$response instanceof Response) {
-            try {
-                $response = $this->Client->get($url);
+        try {
+            $response = $this->Client->get($url);
 
-                if ($this->getConfig('cache') && ($response->isOk() || $response->isRedirect())) {
-                    Cache::write($cacheKey, [$response, (string)$response->getBody()], 'LinkScanner');
-                }
-            } catch (Exception $e) {
-                $response = (new Response())->withStatus(404);
+            if ($this->getConfig('cache') && ($response->isOk() || $response->isRedirect())) {
+                Cache::write($cacheKey, [$response, (string)$response->getBody()], 'LinkScanner');
             }
+        } catch (Exception $e) {
+            $response = (new Response())->withStatus(404);
         }
 
         return $response;
@@ -206,7 +238,7 @@ class LinkScanner implements Serializable
      * Internal method to perform a recursive scan.
      *
      * It recursively repeats the scan for all urls found that have not
-     *  already been scanned.
+     *  already been scanned and until the maximum depth is reached.
      *
      * ### Events
      * This method will trigger some events:
@@ -217,8 +249,8 @@ class LinkScanner implements Serializable
      * @param string $url Url to scan
      * @param string|null $referer Referer of this url
      * @return void
+     * @uses _canBeScanned()
      * @uses _singleScan()
-     * @uses canBeScanned()
      * @uses $alreadyScanned
      * @uses $currentDepth
      * @uses $hostname
@@ -226,7 +258,6 @@ class LinkScanner implements Serializable
     protected function _recursiveScan($url, $referer = null)
     {
         $response = $this->_singleScan($url, $referer);
-
         if (!$response) {
             return;
         }
@@ -245,14 +276,15 @@ class LinkScanner implements Serializable
 
         //Continues scanning for the links found
         foreach ((new BodyParser($response->getBody(), $url))->extractLinks() as $link) {
-            if (!$this->canBeScanned($link)) {
+            if (!$this->_canBeScanned($link)) {
                 continue;
             }
             $this->dispatchEvent('LinkScanner.foundLinkToBeScanned', [$link]);
 
-            //Single scan for external links, recursive scan for internal links
-            $method = is_external_url($link, $this->hostname) ? '_singleScan' : __METHOD__;
-            call_user_func_array([$this, $method], [$link, $url]);
+            //Performs single scan for external links and recursive scan for
+            //  internal links
+            $method = is_external_url($link, $this->hostname) ? '_singleScan' : '_recursiveScan';
+            $this->{$method}($link, $url);
         }
     }
 
@@ -268,41 +300,42 @@ class LinkScanner implements Serializable
      *  - `LinkScanner.foundRedirect`: will be triggered if a redirect is found.
      * @param string $url Url to scan
      * @param string|null $referer Referer of this url
-     * @return \LinkScanner\Http\Client\ScanResponse|null
+     * @return \Cake\Http\Client\Response|null
+     * @uses _canBeScanned()
      * @uses _getResponse()
-     * @uses canBeScanned()
      * @uses $ResultScan
      * @uses $hostname
      */
     protected function _singleScan($url, $referer = null)
     {
         $url = clean_url($url, true, true);
-        if (!$this->canBeScanned($url)) {
+        if (!$this->_canBeScanned($url)) {
             return null;
         }
 
         $this->dispatchEvent('LinkScanner.beforeScanUrl', [$url]);
         $response = $this->_getResponse($url);
+        $location = $response->getHeaderLine('location');
         $this->dispatchEvent('LinkScanner.afterScanUrl', [$response]);
 
         //Follows redirects
         if ($response->isRedirect() && $this->getConfig('followRedirects')) {
-            $location = $response->getHeaderLine('location');
-            if (!$this->canBeScanned($location)) {
+            if (!$this->_canBeScanned($location)) {
                 return null;
             }
 
             $this->dispatchEvent('LinkScanner.foundRedirect', [$location]);
 
-            return call_user_func([$this, __METHOD__], $location);
+            return $this->_singleScan($location);
         }
 
+        //Creates a `ScanEntity` instance with the result and appends it to the
+        //  `ResultScan` instance
         $this->ResultScan = $this->ResultScan->appendItem(new ScanEntity([
             'code' => $response->getStatusCode(),
             'external' => is_external_url($url, $this->hostname),
-            'location' => $response->getHeaderLine('Location'),
             'type' => $response->getHeaderLine('content-type'),
-        ] + compact('url', 'referer')));
+        ] + compact('location', 'url', 'referer')));
 
         return $response;
     }
@@ -395,6 +428,12 @@ class LinkScanner implements Serializable
             __d('link-scanner', 'There is no result to export. Perhaps the scan was not performed?'),
             RuntimeException::class
         );
+
+        if ($this->getConfig('exportOnlyBadResults')) {
+            $this->ResultScan = new ResultScan($this->ResultScan->filter(function (ScanEntity $item) {
+                return $item->get('code') >= 400;
+            }));
+        }
 
         $filename = $this->_getAbsolutePath($filename ?: sprintf('results_%s_%s', $this->hostname, $this->startTime));
         create_file($filename, serialize($this));